Move all OGR unique_* helpers to header
Rather than putting some defintions in .cpp and others .h file, move them all to the header
